Вынер рендеринг карты точек маршрута в контроллер
